requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
pilkit==3.0
redis==5.0.6
rq==1.16.2
sqlparse==0.5.0
tzdata==2024.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git#egg=rq-win